Use the 0BSD license instead of CC-0
CC-0 explicitly does *not* grant a patent use license. It is also way too verbose
and not really designed for software (even though it's still fine to use
it for that).

I've removed the date from the copyright claim since I'd certainly
forget to update it in 2020. The claim isn't needed in the first place.

Closes #10